updated all Thrift bindings using the latest version of the Thrift compi...

+++ src/Thrift/Protocol/TCompactProtocol.cs
@@ -21,10 +21,10 @@
  * details.
  */
 
-using FluentCassandra.Thrift.Transport;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using FluentCassandra.Thrift.Transport;
 
 namespace FluentCassandra.Thrift.Protocol
 {
@@ -317,9 +317,9 @@ namespace FluentCassandra.Thrift.Protocol
         /** 
          * Write a byte. Nothing to see here!
          */
-        public override void WriteByte(byte b)
+        public override void WriteByte(sbyte b)
         {
-            WriteByteDirect(b);
+            WriteByteDirect((byte)b);
         }
 
         /**
@@ -440,7 +440,7 @@ namespace FluentCassandra.Thrift.Protocol
          */
         private ulong longToZigzag(long n)
         {
-            return (ulong)(((ulong)n << 1) ^ ((ulong)n >> 63));
+            return (ulong)(n << 1) ^ (ulong)(n >> 63);
         }
 
         /**
@@ -449,7 +449,7 @@ namespace FluentCassandra.Thrift.Protocol
          */
         private uint intToZigZag(int n)
         {
-            return (uint)(((uint)n << 1) ^ ((uint)n >> 31));
+            return (uint)(n << 1) ^ (uint)(n >> 31);
         }
 
         /**
@@ -477,12 +477,12 @@ namespace FluentCassandra.Thrift.Protocol
    */
         public override TMessage ReadMessageBegin()
         {
-            byte protocolId = ReadByte();
+            byte protocolId = (byte)ReadByte();
             if (protocolId != PROTOCOL_ID)
             {
                 throw new TProtocolException("Expected protocol id " + PROTOCOL_ID.ToString("X") + " but got " + protocolId.ToString("X"));
             }
-            byte versionAndType = ReadByte();
+            byte versionAndType = (byte)ReadByte();
             byte version = (byte)(versionAndType & VERSION_MASK);
             if (version != VERSION)
             {
@@ -520,7 +520,7 @@ namespace FluentCassandra.Thrift.Protocol
          */
         public override TField ReadFieldBegin()
         {
-            byte type = ReadByte();
+            byte type = (byte)ReadByte();
 
             // if it's a stop, then we can return immediately, as the struct is over.
             if (type == Types.STOP)
@@ -565,7 +565,7 @@ namespace FluentCassandra.Thrift.Protocol
         public override TMap ReadMapBegin()
         {
             int size = (int)ReadVarint32();
-            byte keyAndValueType = size == 0 ? (byte)0 : ReadByte();
+            byte keyAndValueType = size == 0 ? (byte)0 : (byte)ReadByte();
             return new TMap(getTType((byte)(keyAndValueType >> 4)), getTType((byte)(keyAndValueType & 0xf)), size);
         }
 
@@ -577,7 +577,7 @@ namespace FluentCassandra.Thrift.Protocol
          */
         public override TList ReadListBegin()
         {
-            byte size_and_type = ReadByte();
+            byte size_and_type = (byte)ReadByte();
             int size = (size_and_type >> 4) & 0x0f;
             if (size == 15)
             {
@@ -618,10 +618,10 @@ namespace FluentCassandra.Thrift.Protocol
         /**
          * Read a single byte off the wire. Nothing interesting here.
          */
-        public override byte ReadByte()
+        public override sbyte ReadByte()
         {
             trans.ReadAll(byteRawBuf, 0, 1);
-            return byteRawBuf[0];
+            return (sbyte)byteRawBuf[0];
         }
 
         /**
@@ -722,7 +722,7 @@ namespace FluentCassandra.Thrift.Protocol
             int shift = 0;
             while (true)
             {
-                byte b = ReadByte();
+                byte b = (byte)ReadByte();
                 result |= (uint)(b & 0x7f) << shift;
                 if ((b & 0x80) != 0x80) break;
                 shift += 7;
@@ -740,7 +740,7 @@ namespace FluentCassandra.Thrift.Protocol
             ulong result = 0;
             while (true)
             {
-                byte b = ReadByte();
+                byte b = (byte)ReadByte();
                 result |= (ulong)(b & 0x7f) << shift;
                 if ((b & 0x80) != 0x80) break;
                 shift += 7;
